LU-11838 osp: atomic64_read() returns s64
Since kernel 4.17 atomic64_read on x86_64 returns s64
instead of long.

Use %llu as the format specifier and explicitly cast it
to unsigned long long.

This patch is back-ported from the following one:
Lustre-commit: dc46952ecd1aa09e738b2de6b1a3076ecbaa740e
Lustre-change: https://review.whamcloud.com/34676

Test-Parameters:trivial
Reviewed-on: https://review.whamcloud.com/35174
Tested-by: Jenkins
Tested-by: Maloo <maloo@whamcloud.com>

+++ lustre/osp/osp_dev.c
@@ -828,8 +828,8 @@ static int osp_sync(const struct lu_env *env, struct dt_device *dev)
 	if (rc != 0)
 		GOTO(out, rc);
 
-	CDEBUG(D_CACHE, "%s: processed %lu\n", d->opd_obd->obd_name,
-	       atomic64_read(&d->opd_sync_processed_recs));
+	CDEBUG(D_CACHE, "%s: processed %llu\n", d->opd_obd->obd_name,
+	       (unsigned long long)atomic64_read(&d->opd_sync_processed_recs));
 
 	while (atomic64_read(&d->opd_sync_processed_recs) < old + recs) {
 		__u64 last = atomic64_read(&d->opd_sync_processed_recs);
